+++ src/main/java/at/hannibal2/skyhanni/data/SkillExperience.kt
@@ -1,5 +1,6 @@
 package at.hannibal2.skyhanni.data
 
+import at.hannibal2.skyhanni.api.SkillApi
 import at.hannibal2.skyhanni.api.event.HandleEvent
 import at.hannibal2.skyhanni.events.ActionBarUpdateEvent
 import at.hannibal2.skyhanni.events.InventoryFullyOpenedEvent
@@ -53,7 +54,7 @@ object SkillExperience {
 
     @HandleEvent
     fun onInventoryFullyOpened(event: InventoryFullyOpenedEvent) {
-        if (event.inventoryName != "Your Skills") return
+        if (!SkillApi.skillMenuDetector.isInside()) return
 
         for ((_, stack) in event.inventoryItems) {
             val name = stack.cleanName
